+++ tests/test_tushare.py
@@ -25,8 +25,9 @@ def main():
     client = TushareClient()
     #df = client.get_realtime_quote("600036.SH")
     #df = client.pro.query('cyq_perf', trade_date='20251231')
-    df = client.pro.query('stk_holdernumber', start_date='20250101', end_date='20251231')
+    #df = client.pro.query('stk_holdernumber', start_date='20250101', end_date='20251231')
     #df = client.pro.query('fund_portfolio', period='20251231')
+    df = client.pro.query('express_vip', ann_date='20260430')
     print(df)
 
 if __name__ == "__main__":